alpha-1.0.001
try to print log of the request.

now can understand: service is executed on server and background, that's
ok, but service port `8090` can't be reach, since there has no outside
access to this port. and based on research that bluehost not provide
port access to the shared IP. so that is it, experiment is success and
over.